# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MobileCRM API Documentation",
        "description": "MobileCRM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/business/hotel/bookings": {
            "get": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "List Hotel Bookings",
                "description": "Returns active or filtered bookings for the business",
                "operationId": "77126e3e0a05f167a845668ad8252463",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Create Booking",
                "operationId": "897e7948f6ce9ae2a5c03b307a9380f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "check_in_date",
                                    "check_out_date",
                                    "room_rate_per_night"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "room_rate_per_night": {
                                        "type": "number"
                                    },
                                    "advance_payment": {
                                        "type": "number"
                                    },
                                    "payment_mode": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/bookings/{id}": {
            "get": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Get Booking Details",
                "operationId": "19cd0d4eeba9295a63482787be20b095",
                "responses": {
                    "200": {
                        "description": "Booking retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Update Booking",
                "operationId": "22f43dc4cbd62b2b0dfc2445801056e9",
                "responses": {
                    "200": {
                        "description": "Booking updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Cancel Booking",
                "operationId": "6d90a0c2e4f692d5cb5aa6fa7a02b80b",
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/bookings/{id}/payments": {
            "post": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Record Payment for Booking",
                "operationId": "05f53fe9b76005d56dbed8c76db2a066",
                "responses": {
                    "200": {
                        "description": "Payment recorded"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/bookings/{id}/check-in": {
            "post": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Check In Booking",
                "operationId": "b10e01ec0831724ddc488173703e4950",
                "responses": {
                    "200": {
                        "description": "Checked in successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/bookings/{id}/check-out": {
            "post": {
                "tags": [
                    "Hotel - Bookings"
                ],
                "summary": "Check Out Booking",
                "operationId": "0b33bd2368b81e9fe9c790050bc54c06",
                "responses": {
                    "200": {
                        "description": "Checked out successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer Token to authenticate",
                "scheme": "bearer"
            }
        }
    },
...
}